+++ AGV_PROJECT_B2/sensor_lib.h
@@ -1,8 +1,8 @@
 #ifndef SENSOR_LIB_H_INCLUDED
 #define SENSOR_LIB_H_INCLUDED
 
-float ultrasoonAfstand_R(void);
-float ultrasoonAfstand_L(void);
+uint16_t ultrasoonAfstand_R(void);
+uint16_t ultrasoonAfstand_L(void);
 
 
 #endif // SENSOR_LIB_H_INCLUDED
